move cluster analyzer into DQM
